requirements.txt
aiohttp==3.8.6
aiohttp-cors==0.7.0
python-dotenv==1.0.0
python-keycloak==3.3.0